enumerations extracted from classes string class changed to class more compare methods
