+++ hqlplus/src/main/java/hqlplus/annotation/Elements.java
@@ -11,5 +11,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 public @interface Elements {
 	String alias();
+
 	String inAliasField();
 }
