+++ hqlplus/src/main/java/hqlplus/annotation/SelectMore.java
@@ -10,6 +10,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface SelectMore {
-	String clazzName();
+    Class clazz();
+
 	String alias();
 }
