+++ hqlplus/src/main/java/hqlplus/enumeration/CompareMethod.java
@@ -1,0 +1,5 @@
+package hqlplus.enumeration;
+
+public enum CompareMethod {
+    EQ, GT, GE, LT, LE, IN, LIKE
+}
